dcm4che/dcm4chee-arc-light#236 : RESTful Service: Delete Patient Permanently - Patient having no studies delete patient from db only if there was no exception caught while deleting its studies, else deletion of patient from db will throw foreign key constraint violation

+++ dcm4chee-arc-delete/src/main/java/org/dcm4chee/arc/delete/impl/DeletionServiceImpl.java
@@ -163,6 +163,7 @@ public class DeletionServiceImpl implements DeletionService {
                 try {
                     sCtx = createStudyDeleteContext(study.getPk(), ctx.getHttpRequest());
                     studyDeleted(sCtx, study, AllowDeleteStudyPermanently.REJECTED);
+                    LOG.info("Successfully delete {} from database", study);
                 } catch (Exception e) {
                     LOG.warn("Failed to delete {} on {}", study, e);
                     ctx.setException(e);
@@ -170,8 +171,10 @@ public class DeletionServiceImpl implements DeletionService {
                 }
             }
         }
-        patientService.deletePatientFromUI(ctx);
-        LOG.info("Successfully delete {} from database", ctx.getPatient());
+        if (ctx.getException() != null) {
+            patientService.deletePatientFromUI(ctx);
+            LOG.info("Successfully delete {} from database", ctx.getPatient());
+        }
     }
 
     private boolean studyDeleted(StudyDeleteContext ctx, Study study, AllowDeleteStudyPermanently allowDeleteStudy) {
